Add good tests
- comments
- ifs
- loops
- structs
- vars

+++ Compiler/tests/good/Variables/01_variables.rs
@@ -1,4 +1,4 @@
 fn main(){
     let x = 6;
-    let y:i8 = 12 + x;
+    let y:i32 = 12 + x;
 }
